pyproject.toml
[project]
name = "consumidor-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.77.0",
    "faiss-cpu>=1.13.2",
    "gradio>=6.9.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "pandas>=2.3.3",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "typing-extensions>=4.15.0",
]
